requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements.txt --resolver=backtracking requirements.in
#
aiohttp==3.8.4
    # via
    #   langchain
    #   openai
aiosignal==1.3.1
    # via aiohttp
altair==4.2.2
    # via streamlit
anyio==3.6.2
    # via
    #   starlette
    #   watchfiles
async-timeout==4.0.2
    # via
    #   aiohttp
    #   langchain
attrs==23.1.0
    # via
    #   aiohttp
    #   jsonschema
backoff==2.2.1
    # via posthog
blinker==1.6.2
    # via streamlit
cachetools==5.3.0
    # via streamlit
certifi==2023.5.7
    # via
    #   clickhouse-connect
    #   requests
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
chromadb==0.3.21
    # via -r requirements.in
click==8.1.3
    # via
    #   nltk
    #   streamlit
    #   uvicorn
clickhouse-connect==0.5.24
    # via chromadb
dataclasses-json==0.5.7
    # via langchain
decorator==5.1.1
    # via validators
duckdb==0.7.1
    # via chromadb
entrypoints==0.4
    # via altair
fastapi==0.95.1
    # via chromadb
filelock==3.12.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
fsspec==2023.5.0
    # via huggingface-hub
gitdb==4.0.10
    # via gitpython
gitpython==3.1.31
    # via streamlit
h11==0.14.0
    # via uvicorn
hnswlib==0.7.0
    # via chromadb
httptools==0.5.0
    # via uvicorn
huggingface-hub==0.14.1
    # via
    #   sentence-transformers
    #   transformers
idna==3.4
    # via
    #   anyio
    #   requests
    #   yarl
importlib-metadata==6.6.0
    # via streamlit
jinja2==3.1.2
    # via
    #   altair
    #   pydeck
    #   torch
joblib==1.2.0
    # via
    #   nltk
    #   scikit-learn
jsonschema==4.17.3
    # via altair
langchain==0.0.147
    # via -r requirements.in
lz4==4.3.2
    # via clickhouse-connect
markdown-it-py==2.2.0
    # via rich
markupsafe==2.1.2
    # via jinja2
marshmallow==3.19.0
    # via
    #   dataclasses-json
    #   marshmallow-enum
marshmallow-enum==1.5.1
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
monotonic==1.6
    # via posthog
mpmath==1.3.0
    # via sympy
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
networkx==3.1
    # via torch
nltk==3.8.1
    # via sentence-transformers
numexpr==2.8.4
    # via langchain
numpy==1.24.3
    # via
    #   altair
    #   chromadb
    #   hnswlib
    #   langchain
    #   numexpr
    #   pandas
    #   pyarrow
    #   pydeck
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   streamlit
    #   torchvision
    #   transformers
openai==0.27.4
    # via -r requirements.in
openapi-schema-pydantic==1.2.4
    # via langchain
packaging==23.1
    # via
    #   huggingface-hub
    #   marshmallow
    #   streamlit
    #   transformers
pandas==2.0.1
    # via
    #   altair
    #   chromadb
    #   streamlit
pillow==9.5.0
    # via
    #   streamlit
    #   torchvision
posthog==3.0.1
    # via chromadb
protobuf==3.20.3
    # via streamlit
pyarrow==12.0.0
    # via streamlit
pydantic==1.10.7
    # via
    #   chromadb
    #   fastapi
    #   langchain
    #   openapi-schema-pydantic
pydeck==0.8.1b0
    # via streamlit
pygments==2.15.1
    # via rich
pympler==1.0.1
    # via streamlit
pyrsistent==0.19.3
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   pandas
    #   posthog
    #   streamlit
python-dotenv==1.0.0
    # via
    #   -r requirements.in
    #   uvicorn
python-frontmatter==1.0.0
    # via -r requirements.in
pytz==2023.3
    # via
    #   clickhouse-connect
    #   pandas
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyyaml==6.0
    # via
    #   huggingface-hub
    #   langchain
    #   python-frontmatter
    #   transformers
    #   uvicorn
regex==2023.5.5
    # via
    #   nltk
    #   transformers
requests==2.30.0
    # via
    #   chromadb
    #   huggingface-hub
    #   langchain
    #   openai
    #   posthog
    #   streamlit
    #   torchvision
    #   transformers
rich==13.3.5
    # via streamlit
scikit-learn==1.2.2
    # via sentence-transformers
scipy==1.10.1
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==2.2.2
    # via chromadb
sentencepiece==0.1.99
    # via sentence-transformers
six==1.16.0
    # via
    #   posthog
    #   python-dateutil
smmap==5.0.0
    # via gitdb
sniffio==1.3.0
    # via anyio
sqlalchemy==1.4.48
    # via langchain
starlette==0.26.1
    # via fastapi
streamlit==1.22.0
    # via -r requirements.in
sympy==1.12
    # via torch
tenacity==8.2.2
    # via
    #   langchain
    #   streamlit
threadpoolctl==3.1.0
    # via scikit-learn
tokenizers==0.13.3
    # via transformers
toml==0.10.2
    # via streamlit
toolz==0.12.0
    # via altair
torch==2.0.1
    # via
    #   sentence-transformers
    #   torchvision
torchvision==0.15.2
    # via sentence-transformers
tornado==6.3.1
    # via streamlit
tqdm==4.65.0
    # via
    #   huggingface-hub
    #   langchain
    #   nltk
    #   openai
    #   sentence-transformers
    #   transformers
transformers==4.29.0
    # via sentence-transformers
typing-extensions==4.5.0
    # via
    #   huggingface-hub
    #   pydantic
    #   starlette
    #   streamlit
    #   torch
    #   typing-inspect
typing-inspect==0.8.0
    # via dataclasses-json
tzdata==2023.3
    # via
    #   pandas
    #   pytz-deprecation-shim
tzlocal==4.3
    # via streamlit
urllib3==2.0.2
    # via
    #   clickhouse-connect
    #   requests
uvicorn[standard]==0.22.0
    # via chromadb
uvloop==0.17.0
    # via uvicorn
validators==0.20.0
    # via streamlit
watchfiles==0.19.0
    # via uvicorn
websockets==11.0.3
    # via uvicorn
yarl==1.9.2
    # via aiohttp
zipp==3.15.0
    # via importlib-metadata
zstandard==0.21.0
    # via clickhouse-connect
